improve map to image
